Drop vault.ui.serviceType: "LoadBalancer"
It is not needed and it adds a requirement to the cluster to have a
proper LadBalancer which is not always the case.

The default in the helm chart is "ClusterIP", so let's leave that
default.

Tested this on an OCP 4.13 SNO cluster (without LB) and the UI is
correctly accessible.